initial working PKCE from a native app

+++ src/OIDCPipeline.Core/OIDCResponseGenerator.cs
@@ -1,5 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using OIDCPipeline.Core.AuthorizationEndpoint;
+﻿using IdentityServer4.Models;
+using IdentityServer4.Stores;
+using Microsoft.AspNetCore.Mvc;
+using OIDCPipeline.Core.Extensions;
+using OIDCPipeline.Core.Validation.Models;
 using System.Collections.Specialized;
 using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
@@ -7,51 +10,64 @@ using System.Threading.Tasks;
 
 namespace OIDCPipeline.Core
 {
-    public class OIDCResponseGenerator : IOIDCResponseGenerator
+    internal class OIDCResponseGenerator : IOIDCResponseGenerator
     {
         private IOIDCPipelineStore _oidcPipelineStore;
+        private IOIDCPipelineAuthorizationCodeStore _authorizationCodeStore;
 
-        public OIDCResponseGenerator(IOIDCPipelineStore oidcPipelineStore)
+        public OIDCResponseGenerator(
+            IOIDCPipelineStore oidcPipelineStore,
+            IOIDCPipelineAuthorizationCodeStore authorizationCodeStore)
         {
             _oidcPipelineStore = oidcPipelineStore;
+            _authorizationCodeStore = authorizationCodeStore;
         }
         public async Task<IActionResult> CreateIdTokenActionResultResponseAsync(
             string key,
             NameValueCollection extras = null, 
             bool delete = true)
         {
+            if (extras == null) extras = new NameValueCollection();
        
             var original = await _oidcPipelineStore.GetOriginalIdTokenRequestAsync(key);
             var downstream = await _oidcPipelineStore.GetDownstreamIdTokenResponseAsync(key);
 
+            if (!string.IsNullOrWhiteSpace(original.Nonce))
+            {
+                extras["nonce"] = original.Nonce;
+            }
+
             var header = new JwtHeader();
             var handler = new JwtSecurityTokenHandler();
-            var idToken = handler.ReadJwtToken(downstream.id_token);
+            var idToken = handler.ReadJwtToken(downstream.IdToken);
             var claims = idToken.Claims.ToList();
             var scope = (from item in claims where item.Type == "scope" select item).FirstOrDefault();
 
-
-
-            var authResponse = new AuthorizeResponse()
+            var authResponse = new AuthorizeResponse();
+            
+            if ( original.CodeChallenge.IsPresent() &&
+                 original.CodeChallengeMethod.IsPresent())
             {
-                Request = new ValidatedAuthorizeRequest
-                {
-                    State = original.state,
-                    RedirectUri = original.redirect_uri,
-                    ResponseMode = original.response_mode
-                },
-                IdentityToken = downstream.id_token,
-                AccessToken = downstream.access_token,
-                Scope = scope?.Value
-               
-            };
-            if (!string.IsNullOrWhiteSpace(original.nonce))
-            {
-                extras["nonce"] = original.nonce;
+                downstream.Extras.Merge(extras);
+
+                // slide out the stored stuff , as we will have a incoming token request with a code
+                await _oidcPipelineStore.StoreDownstreamIdTokenResponseAsync(key, downstream);
+                authResponse.Code = key;
             }
+            else
+            {
+                authResponse.IdentityToken = downstream.IdToken;
+                authResponse.AccessToken = downstream.AccessToken;
+            }
+            authResponse.Request = original;
+            authResponse.Downstream = downstream;
+            authResponse.Scope = scope?.Value;
+
+             
+           
             var authorizeResult = new AuthorizeResult(authResponse,extras);
 
-            if (delete)
+            if (delete && authResponse.Code.IsMissing())
             {
                 await _oidcPipelineStore.DeleteStoredCacheAsync(key);
             }
